extracted cmd and cfg into pkgs
